feat(vue3-package) 自定义清理脚本和命令-> npm run clean

+++ vue3-package/vue3-package-element/clean.ps1
@@ -1,15 +1,27 @@
-     Write-Output "clean project cache..."
+Write-Output "clean project cache..."
 
-     # 删除 node_modules 目录
-     Write-Output "clean node_modules..."
-     Remove-Item -Recurse -Force node_modules
+# 删除 node_modules 目录
+if (Test-Path -Path "node_modules") {
+    Write-Output "clean node_modules..."
+    Remove-Item -Recurse -Force node_modules
+} else {
+    Write-Output "node_modules directory does not exist."
+}
 
-     # 删除 package-lock.json 文件
-     Write-Output "clean package-lock.json..."
-     Remove-Item package-lock.json
+# 删除 package-lock.json 文件
+if (Test-Path -Path "package-lock.json") {
+    Write-Output "clean package-lock.json..."
+    Remove-Item package-lock.json
+} else {
+    Write-Output "package-lock.json file does not exist."
+}
 
-     # 删除 dist 目录（如果存在）
-     Write-Output "clean dist..."
-     Remove-Item -Recurse -Force dist
+# 删除 dist 目录（如果存在）
+if (Test-Path -Path "dist") {
+    Write-Output "clean dist..."
+    Remove-Item -Recurse -Force dist
+} else {
+    Write-Output "dist directory does not exist."
+}
 
-     Write-Output "clean project cache is complete."
+Write-Output "clean project cache is complete."
